+++ backend/package.json
@@ -5,8 +5,10 @@
    "main": "app.js",
    "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1",
-      "start": "node app.ts",
-      "ts-start": "ts-node src/index.ts"
+      "start": "npm run build-ts && npm run serve",
+      "ts-start": "ts-node src/app.ts",
+      "serve": "node ./build/server.js",
+      "build-ts": "tsc"
    },
    "author": "",
    "license": "ISC",
@@ -21,7 +23,9 @@
       "path": "^0.12.7",
       "pg": "^7.10.0",
       "reflect-metadata": "^0.1.13",
-      "typeorm": "0.2.16"
+      "typeorm": "0.2.16",
+      "async":"2.6.2",
+      "request":"2.88.0"
    },
    "devDependencies": {
       "ts-node": "8.1.0",
