+++ backend/tsconfig.json
@@ -12,4 +12,10 @@
       "experimentalDecorators": true,
       "sourceMap": true
     },
+    "include": [
+        "src/**/*"
+    ],
+    "exclude": [
+      "node_modules"
+    ]
 }
